removed win32 platform support
